calculation/pyproject.toml
[tool.poetry]
name = "calculation"
version = "0.1.0"
description = ""
authors = ["Henrik <henrrb@stud.ntnu.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.2"
openpyxl = "^3.0.10"
xlsxwriter = "^3.0.5"
django = "^4.1.5"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
